+++ rspace_client/tests/base_test.py
@@ -16,7 +16,10 @@ RSPACE_URL_ENV = "RSPACE_URL"
 RSPACE_APIKEY_ENV = "RSPACE_API_KEY"
 
 
-def get_datafile(file_or_dir_name: str):
+def get_datafile(file_or_dir_name: str) -> str:
+    """
+    Gets full file path to a file in the 'data' test folder
+    """
     return os.path.join(os.path.dirname(__file__), "data", file_or_dir_name)